+++ pieapp/containers/statusbar/plugin.py
