+++ pieapp/structs/plugins.py
@@ -6,4 +6,4 @@ Built-in app plugins
 class Plugin:
     About = "about"
     Converter = "converter"
-    TestPlugin = "test-plugin"
+    TestPlugin = "testplugin"
