+++ pieapp/wizard/wizard.py
@@ -22,24 +22,23 @@ class LocaleWizardPage(
     LocalesAccessor,
     QtWidgets.QWizardPage
 ):
-    section = Section.User
+    section = Section.Shared
 
     def __init__(self, parent) -> None:
         super().__init__(parent)
 
         self._parent = parent
         self._locales = Config.LOCALES
-        self._cur_locale = self.get_config("locales.locale", Config.DEFAULT_LOCALE)
+        self._cur_locale = self.get_shared_config("locales.locale", Config.DEFAULT_LOCALE, Section.User)
         self._locales_reversed = {v: k for (k, v) in self._locales.items()}
-        self._cur_locale = self._locales.get(self._cur_locale)
 
         self.combo_box = QtWidgets.QComboBox()
         self.combo_box.set_style_sheet("QComboBox{font-size: 12pt;}")
         self.combo_box.insert_item(0, self._locales.pop(self._cur_locale))
         self.combo_box.add_items([self._locales.get(i) for (i, _) in self._locales.items()])
-        self.combo_box.current_index_changed.connect(self.get_result)
+        self.combo_box.currentIndexChanged.connect(self.get_result)
 
-        self.locale_label = QtWidgets.QLabel(self.get_translation("Select locale", Section.Shared))
+        self.locale_label = QtWidgets.QLabel(self.get_translation("Select locale"))
         self.locale_label.set_style_sheet("QLabel{font-size: 25pt; padding-bottom: 20px;}")
 
         layout = QtWidgets.QVBoxLayout()
@@ -74,11 +73,11 @@ class ThemeWizardPage(
         self.combo_box = QtWidgets.QComboBox()
         self.combo_box.set_style_sheet("QComboBox{font-size: 12pt;}")
         self.combo_box.add_items(Managers(SysManager.Assets).get_themes())
-        self.combo_box.current_index_changed.connect(self.get_result)
+        self.combo_box.currentIndexChanged.connect(self.get_result)
 
         self._cur_theme = self.get_shared_config("assets.theme", Managers(SysManager.Assets).get_theme(), Section.User)
 
-        theme_label = self.get_translation("Select theme")
+        theme_label = QtWidgets.QLabel(self.get_translation("Select theme"))
         theme_label.set_style_sheet("QLabel{font-size: 25pt; padding-bottom: 20px;}")
 
         layout = QtWidgets.QVBoxLayout()
@@ -124,7 +123,7 @@ class FfmpegWizardPage(
                 border-radius: 50px;
             }
         """)
-        self.line_edit_button.setIcon(QIcon(
+        self.line_edit_button.set_icon(QIcon(
             Managers(SysManager.Assets).get(Section.Shared, "open-folder.png")
         ))
         self.line_edit_button.set_icon(self.get_asset_icon("open-folder.png"))
@@ -161,7 +160,7 @@ class FfmpegWizardPage(
             )
             self.ffmpeg_path = directory_path
             self.line_edit.set_text(directory_path)
-            self.complete_changed.emit()
+            self.completeChanged.emit()
 
     def get_result(self):
         return self.line_edit.text()
@@ -171,6 +170,7 @@ class FinishWizardPage(
     LocalesAccessor,
     QtWidgets.QWizardPage
 ):
+    section = Section.Shared
 
     def __init__(self, parent) -> None:
         super().__init__(parent)
